support setting footer_links from config
sometimes we want auxillery pages linked from the footer of the site,
this allows setting:

```
footer_links:
  Page Name: /page/link/url
```

It is also desirable to be able to hide pages from
displaying in the main page tree navigation (for example we likely don't
want the same page in the footer AND in the sidebar). To support this we
add the ability to skip rendering a page in the page tree by setting
`hide_in_navigation: true` frontmatter.