Ensure the page can be pinch to zoomed
I've tested in latest versions of iOS and Android.

- iOS v10 onwards ignores disabling pinch to zoom
- Android does not ignore disabling pinch to zoom, although the user can override this.

I have not been able to introduce any negative impact for removing pinch to zoom, the interface scales proportionally and does not obstruct anything. So given that iOS already ignores this I think we should allow this for all users.

This is also a WCAG accessibility requirement.

Fixes https://github.com/alphagov/tech-docs-gem/issues/110